Add files via upload
Changed the formserver.js file so that it will insert the form data directly into the sql database. All that's left is to make sure that when a parent or teacher signs up, their data gets inserted into the right table. For now it only inserts into the parent table. I also fixed a bug in the functions.js file that would add quotations at the end of the final string in the array.