New test case 2 implemented
This resolves issue #60.

Note:
I made changes to reply.js and posting.js, and created a function.js file that takes the post and comment methods of the reply.js and posting.js files and exports them to their respective files. This is done so that I can use another file to retrieve that function to be able to test it.