Delete testscript (#116)
**The following contains all the info on what was no longer needed and disposed from the repository**

* Delete testscript

Deleting this test script as it no longer has a use in our code, and is not used anymore.

* Delete test.js

Deleting this file as it no longer has a use in our code base, and is not used anymore. This is related to issue #98

* Delete cleanup.js

Deleting this file as it no longer has a use in our code base, and is not used anymore. This is related to issue #98

* Delete test1.js

Deleting this file as it no longer has a use in our code base, and is not used anymore. This is related to issue #98

* Delete test1.py

Deleting this file as it no longer has a use in our code base, and is not used anymore. This is related to issue #98

* Delete error.ejs

Deleting this file as it no longer has a use in our code base, and is not used anymore. This is related to issue #98

* Delete homepage.ejs

Deleting this file as it no longer has a use in our code base, and is not used anymore. This is related to issue #98

* Delete teachersignuptest.ejs

Deleting this file as it no longer has a use in our code base, and is not used anymore. This is related to issue #98

* Delete comments.js

Deleting this file as it no longer has a use in our code base, and is not used anymore. This is related to issue #98

* Delete parenthomepage.js

Deleting this file as it no longer has a use in our code base, and is not used anymore. This is related to issue #98

* Delete parentProfile.js

Deleting this file as it no longer has a use in our code base, and is not used anymore. This is related to issue #98

* Delete teacherProfile.js

Deleting this file as it no longer has a use in our code base, and is not used anymore. This is related to issue #98

* Delete users.js

Deleting this file as it no longer has a use in our code base, and is not used anymore. This is related to issue #98

* Delete viewParents.js

Deleting this file as it no longer has a use in our code base, and is not used anymore. This is related to issue #98

* Add files via upload

Made corrections to the file, so that errors are not thrown for files that are no longer there. This is related to issue #98