Fixed nav bar cutting out of screen
The top nav bar where the name used to appear was getting cut out of the screen. Now it is fixed and the page scales properly.
Also made a change to the way the drop down menus open. If the notification drop down menu is open and you click on the second menu next to it, the notification menu will close and vice versa. This was done to prevent overlap of the menus.