Add tagging feature to posts
This commit works towards adding the tagging feature to posts. Issue #148

It completes the first sub-task of dynamically generating a friends list while typing after the "@", and the user can then pick whichever friend they want and it will autocomplete the friend's name.

I also made the connection with the database, which works towards sub-task number 2, and almost completes it, all that needs to be done now is generate the notifications with a SQL statement

As for the third sub-task, no work has been put in yet to port the code to comments just yet.