fixing bug
Fixing bug found through test case #161 .
Essentially what was happening is that the post title variable in the posting function in functions.js was thrown in before a whole bunch of variables, meaning all of its dependencies were not inserting things into the wrong place. I placed the title at the end to prevent most of the dependencies from having to worry about this, and now I changed the file that was dependent on having the title in that position.